requirements.txt
google-generativeai>=0.4.0
python-dotenv>=0.21.0
redis>=4.5.0
pdfplumber>=0.10.0
reportlab>=4.0.0